Major change of architecture .... again. However, this time it actually works ;)
Python -- (wifi) --> SparkCore -- (serial) --> Arduino --(1-wire) --> LED

Lots of network code to handle resets and loss of sync has been incorporated.

Currently, something is off with the thread timers in the transmit_loop, but it shouldn't be too hard to fix (I hope).

Frame rate is ok'ish at slightly over 20fps when scrolling, so performance is in acceptable range (can be optimized).