First beta
Works currently 99%

Made significantly more robust to network interruptions. Will now reset and start listening for new connections after a one second timeout on the link.

Only remaining issue is that it currently requires keepalives to keep link open, as it does not differentiate between a lost connection and a pause in output.